feat: Implement cloudflare deployments delete action
Uses the GitHub JavaScript action as a template for a new action that
can be used to delete Cloudflare Deployments. The new action works by
first paging through all the deployments available through the
Cloudflare API. Then it filters the list of deployments down to those
initiated by GitHub PRs and a matching branch name. It then deletes all
but the latest deployment, which at the moment cannot be deleted.

The action accepts the Cloudflare project name, account id, and token as
well as the branch name to match against as action input variables. It
masks the token to avoid the token being logged.

If the action encounters any errors it will bail out and log an error to
the action console.

The action is unit tested using Jest mocks to mock out the API calls to
Cloudflare. A e2e test is provided, but is skipped during CI as we don't
want to perform actual requests against Cloudflare.

Prettier is used to format code during development.

Signed-off-by: Snorre Magnus Davøen <snorre.magnus.davoen@adventuretech.no>