fix: Force parameter not added for keep === 0
The comparison to check if the force parameter should be supplied used
the string based input argument and not the parsed number when comparing
with the number 0. Thus force would always be false, even for those
cases where keep === 0.

This is now fixed. However, the bug highlights the need to port the
project to TypeScript to more easily catch these bugs.

Signed-off-by: Snorre Magnus Davøen <snorre.magnus.davoen@adventuretech.no>